requirements.txt
# Framework principal y servidor ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Procesamiento de templates
Jinja2==3.1.2

# Base de datos y ORM
SQLAlchemy==2.0.23
alembic==1.12.1
mysqlclient==2.2.0
python-dotenv==1.0.0

# Validación y configuración
pydantic==2.10.2
pydantic-settings==2.1.0

# Seguridad y autenticación
python-jose[cryptography]==3.3.0
passlib>=1.7.4
bcrypt==3.2.2
python-multipart==0.0.6

# Utilidades
python-slugify==8.0.1
email-validator==2.1.0.post1

# Procesamiento de datos
pandas==2.1.3
openpyxl==3.1.2